Fix detekt plugin for AGP 7+
 * Update to AGP 7 and gradle wrapper to 7.1
 * Update hilt to make it work with AGP 7
 * Remove `enableFeaturePreview("GRADLE_METADATA")` which are not needed anymore
 * Remove `javaCompileOptions.annotationProcessorOptions.includeCompileClasspath` which is not supported anymore
 * Add a task to convert SARIF reports into gitlab's JSON format to replace Algirdas' plugin

Algirda's plugin was not working on Java 9 and above because Java 9 now requires modules dependencies as part of project jigsaw.
The libraries must now declare their dependencies and since the plugin was not built against Java 9, this was causing an issue with Gson relying on `java.sql` but not included at runtime.
I couldn't find a way to fix it without either updating Algirda's plugin or making my own Gradle task. I took the latter option as depending on Algirda's code could be an issue if he's not available to fix the issue. Also, the code was fairly easy to duplicate and include into our own codebase.