Revert to download config on gradle invocations
The intended behavior is to always have an up to date version of the
Detekt configuration so that it reflects in the developers' IDE
(through the IDE plugin). Since there are chances Detekt tasks are
only going to be run by CI, only downloading Detekt's rules before
the Detekt tasks would never update the configuration on the
developers' machines.