"Force update" dialog can be launched outside of an Activity.
The information that a forced update is needed, comes into `ApiClient.forceUpdate` method,
which on its own, does not know which activity is running in the foreground. Also, the client apps do not control activities defined in Core, so they wouldn't be able to use the `FragmentManager.showForceUpdate` method in such activities.

This change is for the client apps to be able to show the Force Update dialog
without having access to the current activity.

There is no major difference in behavior of the dialog, except that previously pressing the Back button would finish the activity associated with the dialog; currently, pressing the Back button has no effect — this is to prevent the user from making any other action.

CP-1960