Fix IndexOutOfBoundsException in RecoveryMethodFragment:
In some languages, the word for "terms", when used in a full sentence, may be different
(either conjugation, or simply different letter casing).
In that situation, the `startIndex` was equal to `-1` and the `setSpan` method would crash.

To fix this, we mark the position in a string, where the link should appear.
In case no such annotation is found, the whole link is clickable.

Add comment to `auth_signup_terms_conditions_full` string