Key/User : fix handling of signed key lists
Signed key lists are used by the backend to enable key transparency of address keys.
This commits moves the logic of generating SKLs to the user address level.
It fixes the handling of flags and primariness of keys in SKLs.
We update the model of address and public address entities to store the SKL associated.
We migrate databases to be able to store those SKLs.