Bump version for mail-message to fix a `NoSuchMethodError` crash
The class `GetRecipientPublicAddresses` (which is in mail-message module)
is calling `publicAddressKeyRepository.getPublicAddressOrNull`. At one point,
the third argument of this method was changed from boolean to Source,
so the method signature has changed.

`GetRecipientPublicAddresses` was calling this method only with
two arguments, and the third one was implicit (as it had a default value).

As a result, we didn't get any compilation errors, but the app would crash at runtime.