Add prefetch method to FeatureFlagRepository to allow caching list of flags
This method is needed since "global" flags (aka any flag that is not
directly correlated to a user) are currently not updated through the
event loop. In order to keep them updated in the app while avoiding to
query the API each time a flag is needed, a client can decide to fetch
them all at app launch.

MAILANDR-48