test(human-verification): Ignore `phone` and `email` tests.
For both of these tests, after the user enters the verification code,
the "Verify" web app will send a POST request.
Normally, we intercept all requests, and attach a custom proxy token,
but it's not possible to override a POST request, since `WebResourceRequest`
doesn't provide access to POST body data, so we cannot resend that request on our own.