ci(finish:review): apply same rules to finish:review job as for deploy:review.
This aims to prevent finish:review from running during 'main' branch
pipelines.

By default, that task runs when ATLAS_DEPLOY_ENV is true, which is the
case for all pipelines for this repo (since that variable is defined in
the project settings).
However, as the deploy only runs on MRs, the finish task should do so
too.