feat(event-manager): Added App Standby Buckets support.
The idea is to refresh the events enough often compare to the usage, and the limitation in time and jobs in a window.
For example if the app is only used two or three times a week, refresh the events each 8 hours is fairly enough.
So to get the background interval we choose to refresh the events once per window.

| bucket      | window | minutes | jobs limitation | bg interval | user usage
-----------------------------------------------------------------------------
| active      | 10min  | N/A     | 75              | 30min       | used
| working set | 2h     | 10      | 120             | 2h          | daily
| frequent    | 8h     | 10      | 200             | 8h          | weekly
| rare        | 24h    | 10      | 48              | 24h         | rarely
| restricted  | 24h    | 1       | 1               | 72h         | -